% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_erois.R
\name{add_indirect_energy_to_erois}
\alias{add_indirect_energy_to_erois}
\title{Title}
\usage{
add_indirect_energy_to_erois(
  .tidy_summarised_erois_df,
  .tidy_indirect_energy,
  .tidy_iea_df,
  include_non_energy_uses = TRUE
)
}
\arguments{
\item{include_non_energy_uses}{}
}
\value{

}
\description{
Title
}
