% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idE_calcs.R
\name{add_indirect_energy_useful_erois_by}
\alias{add_indirect_energy_useful_erois_by}
\title{Adds indirect energy to useful stage EROI values with given breakdown}
\usage{
add_indirect_energy_useful_erois_by(
  .tidy_aggregated_erois_by_df,
  .tidy_indirect_energy,
  .tidy_iea_df,
  aggregation_category = "EU_category",
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  e_dot = IEATools::iea_cols$e_dot,
  unit = IEATools::iea_cols$unit,
  product.group = "Product.Group",
  energy.stage = "Energy.stage",
  group.eroi = "Group.eroi",
  total_group_output = "Total_Group_Output",
  indirect_energy_ktoe = "Indirect_Energy_ktoe",
  eroi.method = "Eroi.method",
  type = "Type",
  boundary = "Boundary",
  final_to_useful_eff = "Final_to_useful_efficiency",
  ratio_indirect_energy_per_output = "ratio_indirect_energy_per_output",
  include_non_energy_uses = TRUE
)
}
\arguments{
\item{.tidy_aggregated_erois_by_df}{The aggregated useful stage EROI values, with breakdown, to which indirect energy needs to be added}

\item{.tidy_indirect_energy}{A tidy data frame containing the indirect energy to be added, in ktoe.}

\item{.tidy_iea_df}{The \code{.tidy_iea_df} from which the output per product group is calculated.}

\item{aggregation_category}{The name of the column standing for the breakdown of EROIs. Default is "EU_category".}

\item{country, year, method, energy_type, last_stage, e_dot, unit}{See \code{IEATools::iea_cols}.}

\item{product.group}{The name of the column containing the product group name.
Default is "Product.Group".}

\item{energy.stage}{The name of the column containing the energy stage for the calculation of the EROI.
Default is "Energy.stage".}

\item{group.eroi}{The name of the column containing the EROI values.
Default is "Group.eroi".}

\item{total_group_output}{The name of the column containing the total product group output.
Default is "Total_Group_Output".}

\item{indirect_energy_ktoe}{The name of the column containing the indirect energy data in the input indirect energy data frame.
Default is "Indirect_Energy_ktoe".}

\item{eroi.method}{The name of the column containing the eroi method.
Default is "Eroi.method".}

\item{type}{The name of the column containing the type of eroi calculated.
Default is "Type".}

\item{boundary}{The name of the column containing the boundary for the eroi calculation.
Default is "Boundary".}

\item{final_to_useful_eff}{The name of the temporary column containing the calculated final-to-useful efficiencies.
Default is "Final_to_useful_efficiency".}

\item{ratio_indirect_energy_per_output}{The name of the column containing the ratio of indirect energy to product group output.
Default is "ratio_indirect_energy_per_output".}

\item{include_non_energy_uses}{A boolean stating whether non-energy uses are included in the calculation of the output per product group.}
}
\value{
A data frame containing the EROIs with considered breakdown, with the indirect energy now included.
}
\description{
To add indirect energy, the function calculates the ratio of indirect energy to output per product group,
and then adds it to the energy inputs using the inverse of the EROI. The function starts by calculating the ratio of
indirect energy to final energy output, and then uses the average efficiency, either by end-use or final demand sector depending on the breakdown,
to determine the ratio indirect energy input to useful energy output. The average efficiency is determined as the ratio of useful stage EROI to final stage EROI.
}
\details{
The .tidy_iea_df provided as input should be a representation of the global Energy Conversion Chain, as the ratio of indirect energy to output (primary and final energy stages)
is calculated at the global level, so output per fossil fuel group needs to be calculated at the global level.
}
