% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idE_eroi_helpers.R
\name{calc_fec_from_ff_as_heat_by_group}
\alias{calc_fec_from_ff_as_heat_by_group}
\title{Calculates final energy consumption of heat from fossil fuel origin, by fossil fuel group}
\usage{
calc_fec_from_ff_as_heat_by_group(
  .tidy_iea_df,
  list_use_mats = c(IEATools::psut_cols$Y, IEATools::psut_cols$U_eiou),
  list_supply_mats = c(IEATools::psut_cols$V),
  list_oil_products = IEATools::oil_and_oil_products,
  list_coal_products = IEATools::coal_and_coal_products,
  list_gas_products = IEATools::primary_gas_products,
  exports = IEATools::interface_industries$exports,
  losses = IEATools::tfc_compare_flows$losses,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  matnames = IEATools::mat_meta_cols$matnames,
  product.group = "Product.Group",
  energy.stage = "Energy.stage",
  share = "Share",
  product_without_origin = "product_without_origin"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which final energy consumption of heat by fossil fuel group needs to be determined.}

\item{list_use_mats}{The list of matrices to be used for determining total final energy consumption of heat.
Default is \code{c(IEATools::psut_cols$Y, IEATools::psut_cols$U_eiou)}.}

\item{list_supply_mats}{The list of matrices to be used for determining the shares of supply of heat by fossil fuel group.
Default is \code{c(IEATools::psut_cols$V)}.}

\item{list_oil_products}{The list of oil products to be used when calculating the use shares.
Default is \code{IEATools::oil_and_oil_products}.}

\item{list_coal_products}{The list of coal products to be used when calculating the use shares.
Default is \code{IEATools::coal_and_coal_products}.}

\item{list_gas_products}{The list of gas products to be used when calculating the use shares.
Default is \code{IEATools::primary_gas_products}.}

\item{exports}{The character string identifying exports flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::interface_industries$exports}.}

\item{losses}{The character string identifying losses flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::tfc_compare_flows$losses}.}

\item{country, method, energy_type, last_stage, ledger_side, flow_aggregation_point, year, product, unit, flow, e_dot}{See \code{IEATools::iea_cols}.}

\item{matnames}{The name of the column containing the matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{product.group}{The name of the column containing the product group name.
Default is "Product.Group".}

\item{energy.stage}{The name of the column containing the energy stage.
Default is "Energy.stage".}

\item{share}{The name of the column containing the shares of supply of heat by fossil fuel group.
Default is "Share".}

\item{product_without_origin}{The name of the column containing the product name excluding the origin of the product.
Default is "product_without_origin".}
}
\value{
A tidy data frame containing the total final energy consumption of heat, broken down by the supplying fossil fuel group.
}
\description{
This function calculates the final energy consumption of heat from fossil fuel origin, by fossil fuel group. It does so by determining
the total final consumption of heat (from use matrices) and then applying share of heat supplied by each fossil fuel group (determined from supply matrices).
}
\examples{
ECCTools::tidy_AB_data \%>\% 
   IEATools::add_psut_matnames() \%>\% 
   ECCTools::transform_to_dta() \%>\% 
   calc_fec_from_ff_as_heat_by_group()
}
