% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idE_eroi_helpers.R
\name{calc_fec_from_ff_by_group}
\alias{calc_fec_from_ff_by_group}
\title{Calculates final energy consumption supplied by each fossil fuel group, including in the form of electricity and heat}
\usage{
calc_fec_from_ff_by_group(
  .tidy_iea_df,
  include_non_energy_uses = FALSE,
  list_use_mats = c(IEATools::psut_cols$Y, IEATools::psut_cols$U_eiou),
  list_supply_mats = c(IEATools::psut_cols$V),
  list_oil_products = IEATools::oil_and_oil_products,
  list_coal_products = IEATools::coal_and_coal_products,
  list_gas_products = IEATools::primary_gas_products,
  list_non_energy_flows = IEATools::non_energy_flows,
  exports = IEATools::interface_industries$exports,
  losses = IEATools::tfc_compare_flows$losses,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  matnames = IEATools::mat_meta_cols$matnames,
  product.group = "Product.Group",
  energy.stage = "Energy.stage"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which the final energy consumption supplied by each fossil fuel group needs to be calculated.}

\item{include_non_energy_uses}{A boolean indicating whether non-energy uses are included in the calculation of final energy consumption by fossil fuel group.
Default is "FALSE".}

\item{list_use_mats}{The list of matrices to be used for determining total final energy consumption by fossil fuel group.
Default is \code{c(IEATools::psut_cols$Y, IEATools::psut_cols$U_eiou)}.}

\item{list_supply_mats}{The list of matrices to be used for determining the shares of supply of heat and electricity by fossil fuel group.
Default is \code{c(IEATools::psut_cols$V)}.}

\item{list_oil_products}{The list of oil products to be used when calculating the use shares.
Default is \code{IEATools::oil_and_oil_products}.}

\item{list_coal_products}{The list of coal products to be used when calculating the use shares.
Default is \code{IEATools::coal_and_coal_products}.}

\item{list_gas_products}{The list of gas products to be used when calculating the use shares.
Default is \code{IEATools::primary_gas_products}.}

\item{list_non_energy_flows}{A list identifying non-energy flows.
Default is \code{IEATools::non_energy_flows}.}

\item{exports}{The character string identifying exports flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::interface_industries$exports}.}

\item{losses}{The character string identifying losses flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::tfc_compare_flows$losses}.}

\item{country, method, energy_type, last_stage, year, product, unit, flow, e_dot}{See \code{IEATools::iea_cols}.}

\item{matnames}{The name of the column containing the product group name.
Default is "Product.Group".}

\item{product.group}{The name of the column containing the product group name.
Default is "Product.Group".}

\item{energy.stage}{The name of the column containing the energy stage.
Default is "Energy.stage".}
}
\value{
A tidy data frame containing the final energy consumption supplied by each fossil fuel group, including in the form of heat and electricity.
}
\description{
This function calculates final energy consumption supplied by each fossil fuel group, including in the form of electricity and heat.
It calculates it using the list of use matrices supplied by the user, and the list of supply matrices to determine the share of supply of electricity and heat
by fossil fuel group.
}
\examples{
ECCTools::tidy_AB_data \%>\% 
   IEATools::add_psut_matnames() \%>\% 
   ECCTools::transform_to_dta() \%>\% 
   calc_fec_from_ff_by_group()
}
