% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_supply_use.R
\name{calc_ff_use}
\alias{calc_ff_use}
\title{Calculates total primary fossil fuel energy supply}
\usage{
calc_ff_use(
  .tidy_iea_df,
  include_non_energy_uses = FALSE,
  final_use_mats = c(IEATools::psut_cols$Y, IEATools::psut_cols$U_eiou),
  list_oil_products = IEATools::oil_and_oil_products,
  list_coal_products = IEATools::coal_and_coal_products,
  list_gas_products = IEATools::primary_gas_products,
  list_non_energy_flows = IEATools::non_energy_flows,
  exports = IEATools::interface_industries$exports,
  losses = IEATools::tfc_compare_flows$losses,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  matnames = IEATools::mat_meta_cols$matnames,
  product.group = "Product.Group",
  total_group_use = "Total_Group_Use",
  energy.stage = "Energy.stage",
  product_without_origin = "product_without_origin"
)
}
\arguments{
\item{.tidy_iea_df}{The tidy iea data frame for which total final fuel energy use needs to be calculated.}

\item{include_non_energy_uses}{A boolean defining whether non-energy uses should be included
in the calculation of total final energy use. Default is FALSE.}

\item{final_use_mats}{A list describing from which matrices should total final energy uses be calculated.
Default is \code{c(IEATools::psut_cols$Y, IEATools::psut_cols$U_eiou)}.}

\item{list_oil_products}{A list containing the names of oil products.
Default is \code{IEATools::oil_and_oil_products}.}

\item{list_coal_products}{A list containing the names of coal products.
Default is \code{IEATools::coal_and_coal_products}.}

\item{list_gas_products}{A list containing the names of gas products.
Default is \code{IEATools::primary_gas_products}.}

\item{list_non_energy_flows}{}

\item{exports}{A string identifying Exports flows.
Default is \code{IEATools::interface_industries$exports}.}

\item{losses}{A string identifying Losses flows.
Default is \code{IEATools::tfc_compare_flows$losses}.}

\item{country, method, energy_type, last_stage, year, product, unit, flow, e_dot}{See \code{IEATools::iea_cols}.}

\item{matnames}{The column name of the column having matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{product.group}{The column name of the column defining the fossil fuel group.
Default is "Product.Group".}

\item{total_group_use}{Column name containing total energy use by product group.
Default is "Total_Group_Use".}

\item{energy.stage}{The column name of the column defining the energy stage.
Default is "Energy.stage".}

\item{product_without_origin}{Column name containing the name of the product excluding the country of origin.
Helpful for doing calculations with Global Market Assumption.
Default is "product_without_origin".}

\item{primary_production_mats}{A list containing the names of matrices containing primary production flows.
Default is \code{c(IEATools::psut_cols$V)}.}

\item{list_primary_oil_products}{A list containing the names of primary oil products.
Default is \code{IEATools::primary_oil_products}.}

\item{list_primary_coal_products}{A list containing the names of primary coal products.
Default is \verb{IEATools::primary_coal_products,}.}

\item{list_primary_gas_products}{A list containing the names of primary gas products.
Default is \verb{IEATools::primary_gas_products,}.}

\item{total_group_supply}{Column name containing total energy use by product group.
Default is "Total_Group_Supply".}
}
\value{
A tidy data frame with the total fossil fuel primary energy supply for each observation (country, year, method, last stage).

A tidy data frame with the total final fossil fuel energy use for each observation (country, year, method, last stage).
}
\description{
The function calculates total fossil fuel primary energy supply for each observation (country, year, method, last stage).
By default, it uses flows in the V matrix and selects primary energy products to determine primary energy production flows.
}
\details{
The function can work both on a single country Energy Conversion Chain of Domestic Technology Assumption type,
or with a multi-regional Energy Conversion Chain for instance using the Global Market Assumption. The input data frame
will have to be slightly adapted in this case.
}
\examples{
Calculates total final fossil fuel energy use 

This function calculates total final fossil energy use for each observation (country, year, method, last stage).
By default, the matrices used to calculate final fossil energy use flows are `c(IEATools::psut_cols$Y, IEATools::psut_cols$U_eiou)`.
It excludes losses and exports from the calculation.

The function can work both on a single country Energy Conversion Chain of Domestic Technology Assumption type,
or with a multi-regional Energy Conversion Chain for instance using the Global Market Assumption. The input data frame
will have to be slightly adapted in this case.
Non-energy use flows can be included or excluded from the calculations using the `include_non_energy_uses` boolean.

}
