% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_supply_use.R
\name{calc_primary_products_supply_by_group}
\alias{calc_primary_products_supply_by_group}
\title{Title}
\usage{
calc_primary_products_supply_by_group(
  .tidy_iea_df,
  include_non_energy_uses = FALSE,
  primary_production_mats = c(IEATools::psut_cols$V),
  list_primary_oil_products = IEATools::primary_oil_products,
  list_primary_coal_products = IEATools::primary_coal_products,
  list_primary_gas_products = IEATools::primary_gas_products,
  list_non_energy_flows = IEATools::non_energy_flows,
  exports = IEATools::interface_industries$exports,
  losses = IEATools::tfc_compare_flows$losses,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  matnames = IEATools::mat_meta_cols$matnames,
  product.group = "Product.Group",
  energy.stage = "Energy.stage",
  total_group_use = "Total_Group_Use",
  product_without_origin = "product_without_origin"
)
}
\arguments{
\item{total_group_use}{}
}
\value{

}
\description{
Title
}
