% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idE_eroi_helpers.R
\name{calc_share_elec_supply_by_ff_group}
\alias{calc_share_elec_supply_by_ff_group}
\title{Calculates the share of electricity provided by each fossil fuel group}
\usage{
calc_share_elec_supply_by_ff_group(
  .tidy_iea_df,
  list_supply_mats = c(IEATools::psut_cols$V),
  list_oil_products = IEATools::oil_and_oil_products,
  list_coal_products = IEATools::coal_and_coal_products,
  list_gas_products = IEATools::primary_gas_products,
  exports = IEATools::interface_industries$exports,
  losses = IEATools::tfc_compare_flows$losses,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  matnames = IEATools::mat_meta_cols$matnames,
  product.group = "Product.Group",
  energy.stage = "Energy.stage",
  share = "Share",
  product_without_origin = "product_without_origin"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which the share of electricity provided by each fossil fuel group needs to be calculated.}

\item{list_supply_mats}{The list of the supply matrices to be used for the calculating the shares.
Default is \code{c(IEATools::psut_cols$V)}.}

\item{list_oil_products}{The list of oil products to be used when calculating the use shares.
Default is \code{IEATools::oil_and_oil_products}.}

\item{list_coal_products}{The list of coal products to be used when calculating the use shares.
Default is \code{IEATools::coal_and_coal_products}.}

\item{list_gas_products}{The list of gas products to be used when calculating the use shares.
Default is \code{IEATools::primary_gas_products}.}

\item{exports}{The character string identifying exports flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::interface_industries$exports}.}

\item{losses}{The character string identifying losses flows in the \code{.tidy_iea_df}.
Default is \code{IEATools::tfc_compare_flows$losses}.}

\item{country, method, energy_type, last_stage, year, product, unit, flow, e_dot}{See \code{IEATools::iea_cols}.}

\item{matnames}{The name of the column containing the matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{product.group}{The name of the column containing the product group name.
Default is "Product.Group".}

\item{energy.stage}{The name of the column containing the energy stage for the calculation of the EROI.
Default is "Energy.stage".}

\item{share}{The name of the column containing the shares of heat delivered by each fossil fuel group.}
}
\value{
A tidy data frame containing the shares of heat delivered by each fossil fuel group.
}
\description{
This function calculates, for each country, the share of electricity that has been supplied by each fossil fuel group,
out of the domestically produced electricity. So, shares don't have to add up to unity -
indeed they won't if any electricity is supplied by renewables or nuclear.
}
\details{
Fossil fuel groups used are "All fossil fuels", "Oil and gas products", "Oil products", "Natural gas", and "Coal products",
so they are obviously not exclusive.
}
