% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idE_eroi_helpers.R
\name{calc_share_heat_supply_by_ff_group}
\alias{calc_share_heat_supply_by_ff_group}
\title{Calculates the share of heat provided by each fossil fuel group}
\usage{
calc_share_heat_supply_by_ff_group(
  .tidy_iea_df,
  list_supply_mats = "V",
  list_oil_products = IEATools::oil_and_oil_products,
  list_coal_products = IEATools::coal_and_coal_products,
  list_gas_products = IEATools::primary_gas_products,
  exports = IEATools::interface_industries$exports,
  losses = IEATools::tfc_compare_flows$losses,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  matnames = IEATools::mat_meta_cols$matnames,
  product.group = "Product.Group",
  energy.stage = "Energy.stage",
  share = "Share",
  product_without_origin = "product_without_origin"
)
}
\arguments{
\item{share}{}
}
\value{

}
\description{
This function calculates, for each country, the share of heat that has been supplied by each fossil fuel group,
out of the domestically produced heat. So, shares don't have to add up to unity -
indeed they won't if any heat is supplied by renewables or nuclear.
}
\details{
Fossil fuel groups used are "All fossil fuels", "Oil and gas products", "Oil products", "Natural gas", and "Coal products",
so they are obviously not exclusive.
}
