% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_shares.R
\name{calc_shares_ff_by_group_inc_elec_heat}
\alias{calc_shares_ff_by_group_inc_elec_heat}
\title{Calculates the share of use of each energy product within each fossil fuel group, including electricity and heat uses}
\usage{
calc_shares_ff_by_group_inc_elec_heat(
  .tidy_iea_df,
  include_non_energy_uses = FALSE,
  final_use_mats = c(IEATools::psut_cols$Y, IEATools::psut_cols$U_eiou),
  list_oil_products = IEATools::oil_and_oil_products,
  list_coal_products = IEATools::coal_and_coal_products,
  list_gas_products = IEATools::primary_gas_products,
  product.group = "Product.Group",
  total_product_use = "Total_Product_Use",
  total_group_use = "Total_Group_Use",
  non_energy_uses = "Non_Energy_Uses",
  share = "Share",
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  unit = IEATools::iea_cols$unit,
  e_dot = IEATools::iea_cols$e_dot,
  product = IEATools::iea_cols$product,
  boolean_non_energy_uses = "Boolean_Non_Energy_Uses",
  energy.stage = "Energy.stage",
  product_without_origin = "product_without_origin"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} for which the share of uses should be calculated.}

\item{include_non_energy_uses}{A boolean indicating whether non-energy uses should be included in the calculation of the share of uses.
Default is FALSE.}

\item{final_use_mats}{The list of matrices to be used to calculate the share of uses.
Default is \code{c(IEATools::psut_cols$Y, IEATools::psut_cols$U_eiou)}.}

\item{list_oil_products}{The list of oil products to be used for the shares calculations.
Default is \code{IEATools::oil_and_oil_products}.}

\item{list_coal_products}{The list of coal products to be used for the shares calculations.
Default is \code{IEATools::coal_and_coal_products}.}

\item{list_gas_products}{The list of gas products to be used for the shares calculations.
Default is \code{IEATools::primary_gas_products}.}

\item{product.group}{The name of the column containing the product group name.
Default is "Product.Group".}

\item{total_product_use}{The name of the column containing the total energy use by product.
Default is "Total_Product_Use".}

\item{total_group_use}{The name of the column containing the total energy use by fossil fuel group.
Default is "Total_Group_Use".}

\item{non_energy_uses}{The name of the column stating whether non-energy uses are included in the shares calculations.
Default is "Non_Energy_Uses".}

\item{share}{The name of the column containing the shares of each fossil fuel within the electricity supply.
Default is "Share".}

\item{country, method, energy_type, last_stage, year, unit, e_dot, product}{See \code{IEATools::iea_cols}.}

\item{boolean_non_energy_uses}{The column name of a column containing a boolean stating whether non-energy uses are included in the shares calculations.
Default is "Boolean_Non_Energy_Uses".}

\item{energy.stage}{The name of the column containing the energy stage.
Default is "Energy.stage".}

\item{product_without_origin}{The name of the column containing the product name without the product origin.
Default is "product_without_origin"}
}
\value{
A \code{.tidy_iea_df} with the shares of use of each energy product within each fossil fuel group.
}
\description{
This function calculates the share of use of each energy product within each fossil fuel group,
when including electricity and heat of fossil fuel origin in the calculations.
}
