% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eroi_helpers.R
\name{calc_shares_heat_by_ff_group}
\alias{calc_shares_heat_by_ff_group}
\title{Title}
\usage{
calc_shares_heat_by_ff_group(
  .tidy_iea_df,
  supply_mats_list = c(IEATools::psut_cols$V),
  matnames = IEATools::mat_meta_cols$matnames,
  product.group = "Product.Group",
  share = "Share",
  energy.stage = "Energy.stage",
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  unit = IEATools::iea_cols$unit,
  product = IEATools::iea_cols$product,
  e.dot = IEATools::iea_cols$e_dot,
  product_without_origin = "product_without_origin"
)
}
\arguments{
\item{e.dot}{}
}
\value{

}
\description{
Title
}
