% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_erois.R
\name{extract_tidy_product_erois}
\alias{extract_tidy_product_erois}
\title{Extracts product level EROIs from tidy Input Output matrices}
\usage{
extract_tidy_product_erois(
  .tidy_io_mats,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  matnames = IEATools::mat_meta_cols$matnames,
  matvals = IEATools::mat_meta_cols$matvals,
  eroi_g_p = "eroi_g_p",
  eroi_n_p = "eroi_n_p",
  eroi_g_p_feed = "eroi_g_p_feed",
  eroi_n_p_feed = "eroi_n_p_feed",
  type = "Type",
  boundary = "Boundary",
  eroi = "EROI",
  product = IEATools::iea_cols$product,
  colnames = "colnames",
  rowtypes = "rowtypes",
  coltypes = "coltypes"
)
}
\arguments{
\item{.tidy_io_mats}{The \code{.tidy_io_mats} data frame from which EROIs should be extracted.}

\item{country, method, energy_type, last_stage, year, product}{See \code{IEATools::iea_cols}.}

\item{matnames}{The column name of the column having matrices names.
Default is \code{IEATools::mat_meta_cols$matnames}.}

\item{matvals}{The column name of the column reporting matrices values, once expanded.
Default is \code{IEATools::mat_meta_cols$matvals}.}

\item{eroi_g_p}{The name of the column containing vectors of product-level gross EROIs,
including both energy use for feedstock and EIOU production.
Default is "eroi_g_p".}

\item{eroi_n_p}{The name of the column containing vectors of product-level net EROIs,
including both energy use for feedstock and EIOU production.
Default is "eroi_n_p".}

\item{eroi_g_p_feed}{The name of the column containing vectors of product-level gross EROIs,
including only energy use for feedstock production.
Default is "eroi_g_p_feed".}

\item{eroi_n_p_feed}{The name of the column containing vectors of product-level net EROIs,
including only energy use for feedstock production.
Default is "eroi_g_p_feed".}

\item{type}{The name of the EROI type column (i.e. gross or net EROI).
Default is "Type".}

\item{boundary}{The name of the boundary column.
Default is "Boundary".}

\item{eroi}{The name of the product EROI column in output.
Default is "EROI".}

\item{colnames}{The name of columns when expanding matrices.
Default is "colnames".}

\item{rowtypes}{The name of row types when expanding matrices.
Default is "rowtypes".}

\item{coltypes}{The name of column types when expanding matrices.
Default is "coltypes".}
}
\value{
A data frame reporting calculated EROIs in a tidy format.
}
\description{
The function extracts product level EROIs from tidy Input Output matrices, in a tidy format.
}
\details{
The function can be called after calculating EROIs through the \code{Recca::calc_erois()} function.
The type column indicates whether the calculated EROI is gross or net.
The boundary column indicates whether the EROI includes:
\itemize{
\item Only energy use for feedstock production ("Feedstock");
\item Both energy use for feedstock production and for EIOU production ("All").
}
}
\examples{
# Let's first have a look at the raw data obtained when calculating EROIs:
calculated_erois_raw <- tidy_AB_data \%>\%
 IEATools::prep_psut() \%>\%
 Recca::calc_io_mats() \%>\%
 Recca::calc_E_EIOU() \%>\%
 Recca::calc_erois() \%>\%
 dplyr::glimpse()
# Let's then extract EROIs in a tidy format:
calculated_erois_raw \%>\%
 extract_tidy_product_erois() \%>\%
 print()
}
