% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_shares.R
\name{prepare_gma_for_shares}
\alias{prepare_gma_for_shares}
\title{Prepares tidy iea data frame with Global Market Assumption for shares calculations}
\usage{
prepare_gma_for_shares(
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  product_without_origin = "product_without_origin"
)
}
\arguments{
\item{.tidy_iea_df}{The tidy iea data frame with Global Market Assumption that needs to be prepared for shares calculation.}

\item{country, flow, product}{See \code{IEATools::iea_cols}.}

\item{product_without_origin}{The name of the column containing the product names without the country of production.}
}
\value{
A tidy data frame with Global Market Assumption ready for shares calculations.
}
\description{
This function prepares a tidy iea data frame following the Global Market Assumption for the calculation
of shares of product use within each fossil fuel group. To do this, the country of location is removed from each flow,
and is relocated in the country column, and a column of product without origin is added to the data frame.
}
