% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidy_AB_data_gma}
\alias{tidy_AB_data_gma}
\title{Tidy data frame for the AB data with the Global Market Assumption}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 94 rows and 12 columns.
}
\usage{
tidy_AB_data_gma
}
\description{
A tidy_iea_df containing the representation of the dummy AB Energy Conversion Chain
using the Global Market Assumption
}
\examples{
tidy_AB_data_gma
}
\keyword{datasets}
