% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidy_AB_erois_gma}
\alias{tidy_AB_erois_gma}
\title{Tidy data frame containing EROIs for AB using the Global Market Assumption}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60 rows and 10 columns.
}
\usage{
tidy_AB_erois_gma
}
\description{
Tidy data frame containing EROI values calculated for the AB Energy Conversion Chain
using the Global Market Assumption
}
\examples{
tidy_AB_erois_gma
}
\keyword{datasets}
