% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idE_calcs.R
\name{add_indirect_energy_to_erois}
\alias{add_indirect_energy_to_erois}
\title{Title}
\usage{
add_indirect_energy_to_erois(
  .tidy_summarised_erois_df,
  .tidy_indirect_energy,
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  e_dot = IEATools::iea_cols$e_dot,
  unit = IEATools::iea_cols$unit,
  product.group = "Product.Group",
  energy.stage = "Energy.stage",
  total_group_output = "Total_Group_Output",
  indirect_energy_ktoe = "Indirect_Energy_ktoe",
  type = "Type",
  boundary = "Boundary",
  include_non_energy_uses = TRUE
)
}
\arguments{
\item{.tidy_summarised_erois_df}{}

\item{.tidy_indirect_energy}{}

\item{.tidy_iea_df}{}

\item{country}{}

\item{year}{}

\item{method}{}

\item{energy_type}{}

\item{last_stage}{}

\item{e_dot}{}

\item{unit}{}

\item{product.group}{}

\item{energy.stage}{}

\item{total_group_output}{}

\item{indirect_energy_ktoe}{}

\item{type}{}

\item{boundary}{}

\item{include_non_energy_uses}{}
}
\value{

}
\description{
Title
}
