% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_erois.R
\name{aggregate_primary_stage_erois}
\alias{aggregate_primary_stage_erois}
\title{Title}
\usage{
aggregate_primary_stage_erois(
  .tidy_erois_df,
  .tidy_iea_df,
  primary_production_mats = c(IEATools::psut_cols$V),
  list_primary_oil_products = IEATools::primary_oil_products,
  list_primary_coal_products = IEATools::primary_coal_products,
  list_primary_gas_products = IEATools::primary_gas_products,
  product.group = "Product.Group",
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  eroi.method = "Eroi.method",
  type = "Type",
  boundary = "Boundary",
  share = "Share",
  eroi = "EROI",
  group.eroi = "Group.eroi",
  energy.stage = "Energy.stage",
  product_without_origin = "product_without_origin",
  eroi_calc_method = "dta"
)
}
\arguments{
\item{.tidy_erois_df}{}

\item{.tidy_iea_df}{}

\item{primary_production_mats}{}

\item{list_primary_oil_products}{}

\item{list_primary_coal_products}{}

\item{list_primary_gas_products}{}

\item{product.group}{}

\item{country}{}

\item{method}{}

\item{energy_type}{}

\item{last_stage}{}

\item{year}{}

\item{product}{}

\item{eroi.method}{}

\item{type}{}

\item{boundary}{}

\item{share}{}

\item{eroi}{}

\item{group.eroi}{}

\item{energy.stage}{}

\item{product_without_origin}{}
}
\value{

}
\description{
Title
}
