% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_erois.R
\name{aggregate_useful_stage_erois}
\alias{aggregate_useful_stage_erois}
\title{Title}
\usage{
aggregate_useful_stage_erois(
  .tidy_erois_df,
  .tidy_iea_df,
  include_non_energy_uses = FALSE,
  final_use_mats = c(IEATools::psut_cols$Y, IEATools::psut_cols$U_eiou),
  list_primary_oil_products = IEATools::primary_oil_products,
  list_primary_coal_products = IEATools::primary_coal_products,
  list_primary_gas_products = IEATools::primary_gas_products,
  list_oil_products = IEATools::oil_and_oil_products,
  list_coal_products = IEATools::coal_and_coal_products,
  list_gas_products = IEATools::primary_gas_products,
  product.group = "Product.Group",
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  non_energy_uses = "Non_Energy_Uses",
  eroi.method = "Eroi.method",
  type = "Type",
  boundary = "Boundary",
  share = "Share",
  useful_stage_eroi = "Useful_Stage_EROI",
  group.eroi = "Group.eroi",
  energy.stage = "Energy.stage",
  product_without_origin = "product_without_origin",
  eroi_calc_method = c("dta", "gma")
)
}
\arguments{
\item{product_without_origin}{}
}
\value{

}
\description{
Title
}
