% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_stage.R
\name{calc_avg_efficiency_by_ff_group}
\alias{calc_avg_efficiency_by_ff_group}
\title{Title}
\usage{
calc_avg_efficiency_by_ff_group(
  .tidy_efficiencies_df,
  .tidy_iea_df,
  include_non_energy_uses = FALSE,
  final_use_mats = c(IEATools::psut_cols$Y, IEATools::psut_cols$U_eiou),
  list_primary_oil_products = IEATools::primary_oil_products,
  list_primary_coal_products = IEATools::primary_coal_products,
  list_primary_gas_products = IEATools::primary_gas_products,
  list_oil_products = IEATools::oil_and_oil_products,
  list_coal_products = IEATools::coal_and_coal_products,
  list_gas_products = IEATools::primary_gas_products,
  product.group = "Product.Group",
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  non_energy_uses = "Non_Energy_Uses",
  eroi.method = "Eroi.method",
  type = "Type",
  boundary = "Boundary",
  share = "Share",
  useful_stage_eroi = "Useful_Stage_EROI",
  group.eroi = "Group.eroi",
  energy.stage = "Energy.stage",
  product_without_origin = "product_without_origin",
  average_efficiency = "Average_Efficiency_Col",
  aggregated_efficiency = "Aggregated_Efficiency",
  calc_method = c("dta", "gma")
)
}
\arguments{
\item{.tidy_efficiencies_df}{}

\item{.tidy_iea_df}{}

\item{include_non_energy_uses}{}

\item{final_use_mats}{}

\item{list_primary_oil_products}{}

\item{list_primary_coal_products}{}

\item{list_primary_gas_products}{}

\item{list_oil_products}{}

\item{list_coal_products}{}

\item{list_gas_products}{}

\item{product.group}{}

\item{country}{}

\item{method}{}

\item{energy_type}{}

\item{last_stage}{}

\item{year}{}

\item{product}{}

\item{non_energy_uses}{}

\item{eroi.method}{}

\item{type}{}

\item{boundary}{}

\item{share}{}

\item{useful_stage_eroi}{}

\item{group.eroi}{}

\item{energy.stage}{}

\item{product_without_origin}{}

\item{calc_method}{}
}
\value{

}
\description{
Title
}
