% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idE_eroi_helpers.R
\name{calc_share_elec_supply_by_ff_group}
\alias{calc_share_elec_supply_by_ff_group}
\title{Title}
\usage{
calc_share_elec_supply_by_ff_group(
  .tidy_iea_df,
  list_supply_mats = "V",
  list_oil_products = IEATools::oil_and_oil_products,
  list_coal_products = IEATools::coal_and_coal_products,
  list_gas_products = IEATools::primary_gas_products,
  exports = IEATools::interface_industries$exports,
  losses = IEATools::tfc_compare_flows$losses,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  matnames = IEATools::mat_meta_cols$matnames,
  product.group = "Product.Group",
  energy.stage = "Energy.stage",
  share = "Share",
  product_without_origin = "product_without_origin"
)
}
\arguments{
\item{share}{}
}
\value{

}
\description{
Title
}
