% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eroi_helpers.R
\name{calc_total_use_by_product}
\alias{calc_total_use_by_product}
\title{Title}
\usage{
calc_total_use_by_product(
  .tidy_iea_df,
  include_non_energy_uses = FALSE,
  total_use_mats = c(IEATools::psut_cols$Y, IEATools::psut_cols$U_eiou),
  list_oil_products = IEATools::oil_and_oil_products,
  list_coal_products = IEATools::coal_and_coal_products,
  list_gas_products = IEATools::primary_gas_products,
  list_non_energy_flows = IEATools::non_energy_flows,
  exports = IEATools::interface_industries$exports,
  losses = IEATools::tfc_compare_flows$losses,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  matnames = IEATools::mat_meta_cols$matnames,
  total_product_use = "Total_Product_Use",
  product_without_origin = "product_without_origin"
)
}
\arguments{
\item{total_product_use}{}
}
\value{

}
\description{
Title
}
