% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_stage.R
\name{push_to_useful_erois}
\alias{push_to_useful_erois}
\title{Title}
\usage{
push_to_useful_erois(
  .tidy_io_erois,
  tidy_FU_efficiencies,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  product_without_origin = "product_without_origin",
  average_efficiency = "Average_Efficiency_Col",
  eroi = "EROI",
  useful_stage_eroi = "Useful_Stage_EROI",
  eroi_calc_method = c("dta", "gma")
)
}
\arguments{
\item{.tidy_io_erois}{}

\item{tidy_FU_efficiencies}{}

\item{country}{}

\item{method}{}

\item{energy_type}{}

\item{year}{}

\item{product}{}

\item{product_without_origin}{}

\item{average_efficiency}{}

\item{eroi}{}

\item{useful_stage_eroi}{}

\item{eroi_calc_method}{}
}
\value{

}
\description{
Title
}
